# [cluster-test] Fixed wrong divide to mod operation to avoid out of index