🥇 

## Motivation

(Write your motivation for proposed changes here.)

### Have you read the [Contributing Guidelines on pull requests](https://github.com/libra/libra/blob/master/CONTRIBUTING.md#pull-requests)?

(Write your answer here.)

## Test Plan

Tested with case: instances length is less than seed account, 3 nodes, 9 seeds account.
./target/debug/cluster-test --mint-file="./mint.key" --swarm --emit-tx --chain-id=DEVNET --peers=mgorven.libra.aws.hlw3truzy4ls.com:80:80,34.66.49.119:80:80,40.76.153.93:80:80, --duration=60 --accounts-per-client=12 --workers-per-ac=64


- Before:
```
2020-10-15T18:37:26.315416Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:234 Will use 64 workers per AC with total 192 AC clients
2020-10-15T18:37:26.315503Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:246 Will create 12 accounts_per_client with total 2304 accounts
2020-10-15T18:37:26.316005Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:376 Creating and minting faucet account
2020-10-15T18:37:28.942065Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:394 DD account current balances are 9223371962928775807, requested 2304000000 coins
2020-10-15T18:37:28.942323Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:411 Creating and minting seeds accounts
2020-10-15T18:37:32.802040Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:422 Completed creating seed accounts
2020-10-15T18:37:37.947238Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:480 Completed minting seed accounts
2020-10-15T18:37:37.947292Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:481 Minting additional 2304 accounts
thread 'main' panicked at 'index out of bounds: the len is 3 but the index is 3', testsuite/cluster-test/src/tx_emitter.rs:491:32
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```

- After
```
2020-10-15T18:41:20.355465Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:234 Will use 64 workers per AC with total 192 AC clients
2020-10-15T18:41:20.355533Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:246 Will create 12 accounts_per_client with total 2304 accounts
2020-10-15T18:41:20.355902Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:375 Creating and minting faucet account
2020-10-15T18:41:22.947833Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:393 DD account current balances are 9223371958308775807, requested 2304000000 coins
2020-10-15T18:41:22.948094Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:410 Creating and minting seeds accounts
2020-10-15T18:41:25.697732Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:421 Completed creating seed accounts
2020-10-15T18:41:31.556464Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:471 Completed minting seed accounts
2020-10-15T18:41:31.556504Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:472 Minting additional 2304 accounts
2020-10-15T18:42:51.550799Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:510 Mint is done
2020-10-15T18:42:51.579700Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:280 Tx emitter workers started
2020-10-15T18:43:01.581592Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:539 submitted: 406 txn/s, committed: 216 txn/s, expired: 0 txn/s, latency: 4444 ms, p99 latency: 6050 ms
2020-10-15T18:43:11.583044Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:539 submitted: 379 txn/s, committed: 448 txn/s, expired: 0 txn/s, latency: 4219 ms, p99 latency: 6700 ms
2020-10-15T18:43:21.584783Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:539 submitted: 371 txn/s, committed: 327 txn/s, expired: 0 txn/s, latency: 4174 ms, p99 latency: 5050 ms
2020-10-15T18:43:31.586259Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:539 submitted: 421 txn/s, committed: 391 txn/s, expired: 0 txn/s, latency: 4363 ms, p99 latency: 5450 ms
2020-10-15T18:43:41.587957Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:539 submitted: 408 txn/s, committed: 440 txn/s, expired: 0 txn/s, latency: 4052 ms, p99 latency: 4850 ms
2020-10-15T18:43:51.589695Z [main] INFO testsuite/cluster-test/src/tx_emitter.rs:539 submitted: 413 txn/s, committed: 417 txn/s, expired: 0 txn/s, latency: 4076 ms, p99 latency: 4950 ms
Total stats: submitted: 24720, committed: 24720, expired: 0
Average rate: submitted: 412 txn/s, committed: 412 txn/s, expired: 0 txn/s, latency: 4154 ms, p99 latency: 5550 ms
```

## Related PRs

(If this PR adds or changes functionality, please take some time to update the docs at https://github.com/libra/libra/developers.libra.org, and link to your PR here.)